Version 1.4, Custom PCB & Galvanic Isolation

+++ lib/VCS_Actuators/vcs_throttle.h
@@ -4,11 +4,14 @@
 #include <Arduino.h>
 #include "vcs_pins.h"
 #include "vcs_constants.h"
+#include "vcs_state_machine.h"
 #include <QuickPID.h>
 
-// Variables exposed for telemetry and state machine
+
+// Variables exposed for telemetry and state machine.
+// Changed current_pwm_duty to uint16_t to support 10-bit (0-1023) resolution without overflowing.
 extern uint16_t current_throttle_adc;
-extern uint8_t current_pwm_duty;
+extern uint16_t current_pwm_duty;
 
 void initThrottle();
 void updateThrottle(float measured_rpm, float target_rpm);
